backend/requirements.txt
psycopg2-binary==2.9.9
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
bcrypt==4.1.2
python-dotenv==1.0.0

